Revert "MWPW-137866: Web component-based Merch Card in Milo (#1504)"
This reverts commit 082c30d2da9289caae19fceaf1586193cf3c2ca9.
