+++ libs/blocks/merch-card/merch-card.js
@@ -1,188 +1,184 @@
 /* eslint-disable prefer-destructuring */
 import { decorateButtons, decorateBlockHrs } from '../../utils/decorate.js';
-import { getConfig, createTag } from '../../utils/utils.js';
-import { getUpFromSectionMetadata } from '../card/cardUtils.js';
-import { decorateLinkAnalytics } from '../../martech/attributes.js';
+import { loadStyle, getConfig, createTag } from '../../utils/utils.js';
+import { addBackgroundImg, addWrapper, addFooter } from '../card/cardUtils.js';
 import { replaceKey } from '../../features/placeholders.js';
-import '../../deps/commerce.js';
-import '../../deps/merch-card.js';
 
-const cardTypes = ['segment', 'special-offers', 'plans', 'catalog', 'product'];
+const { miloLibs, codeRoot } = getConfig();
+const base = miloLibs || codeRoot;
 
-const textStyles = {
-  H5: 'detail-m',
-  H4: 'body-xxs',
-  H3: 'heading-xs',
-  H2: 'heading-m',
-  H1: 'heading-l',
+const SEGMENT_BLADE = 'SegmentBlade';
+const SPECIAL_OFFERS = 'SpecialOffers';
+const PLANS_CARD = 'PlansCard';
+const cardTypes = {
+  segment: SEGMENT_BLADE,
+  'special-offers': SPECIAL_OFFERS,
+  plans: PLANS_CARD,
+};
+const getPodType = (styles) => {
+  const authoredType = styles?.find((style) => style in cardTypes);
+  return cardTypes[authoredType] || SEGMENT_BLADE;
 };
 
-const getPodType = (styles) => styles?.find((style) => cardTypes.includes(style));
+const decorateFooter = (el, altCtaMetaData, cardType) => {
+  const cardFooter = el.querySelector('.consonant-CardFooter');
+  const decorateWithSecureTransactionSign = () => {
+    const secureTransactionWrapper = createTag('div', { class: 'secure-transaction-wrapper' });
+    const label = createTag('span', { class: 'secure-transaction-label' });
+    const secureElement = createTag('span', { class: 'secure-transaction-icon' });
+    secureTransactionWrapper.append(secureElement, label);
+    replaceKey('secure-transaction', getConfig()).then((replacedKey) => {
+      label.textContent = replacedKey;
+    });
+    return secureTransactionWrapper;
+  };
 
-const checkBoxLabel = (ctas, altCtaMetaData) => {
-  const altCtaRegex = /href=".*"/;
-  if (!altCtaRegex.test(altCtaMetaData[1]?.innerHTML)) return null;
-  const allButtons = ctas.querySelectorAll('a');
-  const originalCtaButton = allButtons[allButtons.length - 1];
-  const altCtaButtonData = altCtaMetaData[1];
-  const altCtaButton = createTag('a', { class: [...originalCtaButton.classList, 'alt-cta', 'button--inactive'].join(' ') }, altCtaButtonData.textContent);
-  originalCtaButton.classList.add('active');
-  decorateButtons(altCtaButton);
-  ctas.appendChild(altCtaButton);
-  return altCtaMetaData[0].textContent;
+  const createCheckbox = (checkBoxText) => {
+    const container = createTag('label', { class: 'checkbox-container' });
+    const input = createTag('input', { id: 'alt-cta', type: 'checkbox' });
+    const checkmark = createTag('span', { class: 'checkmark' });
+    const label = createTag('span', { class: 'checkbox-label' }, checkBoxText.innerHTML);
+    container.append(input, checkmark, label);
+    return container;
+  };
+
+  const createSecureSign = () => {
+    const cardFooterRow = el.querySelector('.consonant-CardFooter-row');
+    const standardWrapper = createTag('div', { class: 'standard-wrapper' });
+    const secureTransactionWrapper = decorateWithSecureTransactionSign();
+    standardWrapper.append(secureTransactionWrapper, cardFooterRow);
+    cardFooter?.append(standardWrapper);
+  };
+
+  const decorateAlternativeCta = () => {
+    const altCtaRegex = /href=".*"/;
+    if (!altCtaRegex.test(altCtaMetaData[1]?.innerHTML)) return;
+    const cardFooterRow = el.querySelector('.consonant-CardFooter-row');
+    const originalCtaButton = cardFooterRow.querySelector('.consonant-CardFooter-cell--right');
+    const checkboxContainer = createCheckbox(altCtaMetaData[0]);
+    const altCtaButtonData = altCtaMetaData[1];
+    decorateButtons(altCtaButtonData);
+    const altCtaButton = createTag('div', { class: originalCtaButton.classList }, altCtaButtonData.innerHTML);
+    altCtaButton.classList.add('button--inactive');
+    checkboxContainer.querySelector('input[type="checkbox"]').addEventListener('change', ({ target: { checked } }) => {
+      originalCtaButton.classList.toggle('button--inactive', checked);
+      altCtaButton.classList.toggle('button--inactive', !checked);
+    });
+    cardFooterRow.append(altCtaButton);
+    cardFooter.prepend(checkboxContainer);
+    altCtaMetaData[0].parentNode.remove();
+  };
+  if (altCtaMetaData !== null) decorateAlternativeCta();
+  if (el.classList.contains('secure')) createSecureSign();
+  cardFooter.querySelectorAll('.consonant-CardFooter-cell').forEach((cell) => cell.classList.add(`consonant-${cardType}-cell`));
 };
 
-const isHeadingTag = (tagName) => /^H[1-5]$/.test(tagName);
-const isParagraphTag = (tagName) => tagName === 'P';
-
-const createAndAppendTag = (tagName, attributes, content, parent) => {
-  const newTag = createTag(tagName, attributes, content);
-  parent.append(newTag);
-  return newTag;
-};
-
-const parseContent = (el, altCta, cardType, merchCard) => {
+const addInner = (el, altCta, cardType, merchCard) => {
   const innerElements = [...el.querySelectorAll('h1, h2, h3, h4, h5, h6, p, ul')];
-  const bodySlot = createTag('div', { slot: 'body-xs' });
+  const styles = [...el.classList];
+  const merch = styles.includes('merch-card');
+  const pElement = merch && el.querySelector(':scope > div > div > p:last-of-type');
+  const links = pElement ? pElement.querySelectorAll('a') : el.querySelectorAll('a');
+  const list = el.querySelector('ul');
+
+  const inner = el.querySelector(':scope > div:not([class])');
+  inner.classList.add(`consonant-${cardType}-inner`);
 
   innerElements.forEach((element) => {
-    const { tagName } = element;
-    if (isHeadingTag(tagName)) {
-      createAndAppendTag(tagName, { slot: textStyles[tagName] }, element.innerHTML, merchCard);
-      return;
-    }
-    if (isParagraphTag(tagName)) {
-      bodySlot.append(element);
+    if (element.tagName.match(/^H[1-6]$/)) element.classList.add(`consonant-${cardType}-title`);
+    if (element.tagName.match(/^P$/)) element.classList.add(`consonant-${cardType}-description`);
+    if (element.tagName.match(/^UL$/)) {
+      list.classList.add(`consonant-${cardType}-list`);
+      list.querySelectorAll('li').forEach((li) => li.classList.add(`consonant-${cardType}-list-item`));
     }
   });
-  merchCard.append(bodySlot);
+
+  inner.append(...innerElements);
+  addFooter(links, inner, cardType !== PLANS_CARD);
+  decorateFooter(el, altCta, cardType);
+  merchCard.append(inner);
 };
 
-const returnRibbonStyle = (ribbonMetadata) => {
+const decorateRibbon = (el, ribbonMetadata, cardType) => {
   const ribbonStyleRegex = /^#[0-9a-fA-F]+, #[0-9a-fA-F]+$/;
-  if (!ribbonStyleRegex.test(ribbonMetadata[0]?.innerText)) return null;
-  const style = ribbonMetadata[0].innerText;
-  const badgeBackgroundColor = style.split(',')[0].trim();
-  const badgeColor = style.split(',')[1].trim();
+  if (!ribbonStyleRegex.test(ribbonMetadata[0]?.innerText)) return;
+  const ribbonStyle = ribbonMetadata[0].innerText;
+  const [ribbonBgColor, ribbonTextColor] = ribbonStyle.split(', ');
   const ribbonWrapper = ribbonMetadata[0].parentNode;
-  const badgeText = ribbonMetadata[1].innerText;
+  const ribbon = ribbonMetadata[1];
+  ribbon.classList.add(`consonant-${cardType}-ribbon`);
+  const borderStyle = `1px solid ${ribbonBgColor}`;
+  if (el.classList.contains('evergreen')) {
+    ribbon.style.border = borderStyle;
+    ribbon.style.borderRight = 'none';
+  } else {
+    ribbon.style.backgroundColor = ribbonBgColor;
+    el.style.border = borderStyle;
+  }
+  ribbon.style.color = ribbonTextColor;
+  const picture = el.querySelector(`.consonant-${cardType}-img`);
+  if (picture) {
+    picture.insertAdjacentElement('afterend', ribbon);
+  } else {
+    el.insertAdjacentElement('beforeend', ribbon);
+  }
   ribbonWrapper.remove();
-  return { badgeBackgroundColor, badgeColor, badgeText };
 };
 
-const getActionMenuContent = (el, ribbonMetadata) => {
-  const index = (ribbonMetadata !== null) ? 1 : 0;
-  const expectedChildren = ribbonMetadata !== null ? 3 : 2;
-  if (el.childElementCount !== expectedChildren) {
-    return null;
-  }
-  const actionMenuContentWrapper = el.children[index];
-  const actionMenuContent = actionMenuContentWrapper.children[0];
-  actionMenuContentWrapper.remove();
-  return actionMenuContent;
+const decorateIcon = (el, icons, cardType) => {
+  if (!icons) return;
+  const inner = el.querySelector(`.consonant-${cardType}-inner`);
+  const iconWrapper = createTag('div', { class: `consonant-${cardType}-iconWrapper` });
+  icons?.forEach((icon) => {
+    const url = icon.querySelector('img').src;
+    const iconDiv = createTag('div', { class: 'consonant-MerchCard-ProductIcon', style: `background-image: url(${url})` });
+    iconWrapper.appendChild(iconDiv);
+    icon.parentNode?.remove();
+  });
+  inner.prepend(iconWrapper);
 };
-
-function getMerchCardRows(rows, ribbonMetadata, cardType, actionMenuContent) {
-  if (cardType === 'catalog') {
-    const index = ribbonMetadata === null ? 0 : 1;
-    return actionMenuContent !== null ? rows[index + 1] : rows[index];
-  }
-  return rows[ribbonMetadata === null ? 0 : 1];
-}
 
 const init = (el) => {
-  let section = el.closest('.section');
-  const upClass = getUpFromSectionMetadata(section);
-  if (upClass) {
-    import('./legacy-merch-card.js')
-      .then((module) => {
-        module.initLegacyMerchCard(el);
-      });
-  } else {
-    if (section.parentElement.classList.contains('fragment')) {
-      const fragment = section.parentElement;
-      const fragmentParent = fragment.parentElement;
-      section.style.display = 'contents';
-      fragment.style.display = 'contents';
-      fragmentParent.style.display = 'contents';
-      section = fragmentParent.parentElement;
-    }
-    const headings = el.querySelectorAll('h1, h2, h3, h4, h5, h6');
-    decorateLinkAnalytics(el, headings);
-    const images = el.querySelectorAll('picture');
-    let image;
-    const icons = [];
-    const rows = el.querySelectorAll(':scope > *');
-    const styles = [...el.classList];
-    const cardType = getPodType(styles);
-    const ribbonMetadata = rows[0].children?.length === 2 ? rows[0].children : null;
-    const actionMenuContent = cardType === 'catalog' ? getActionMenuContent(el, ribbonMetadata) : null;
-    const row = getMerchCardRows(rows, ribbonMetadata, cardType, actionMenuContent);
-    const altCta = rows[rows.length - 1].children?.length === 2
-      ? rows[rows.length - 1].children : null;
-    const allPElems = row.querySelectorAll('p');
-    const ctas = allPElems[allPElems.length - 1];
-    images.forEach((img) => {
-      const imgNode = img.querySelector('img');
-      const { width, height } = imgNode;
-      const isSquare = Math.abs(width - height) <= 10;
-      if (img) {
-        if (isSquare) {
-          icons.push(img);
-        } else {
-          image = img;
-        }
-      }
-    });
-
-    const merchCard = createTag('merch-card', { class: el.className, variant: cardType });
-    if (ribbonMetadata !== null) {
-      const badge = returnRibbonStyle(ribbonMetadata);
-      if (badge !== null) {
-        merchCard.setAttribute('badge-background-color', badge.badgeBackgroundColor);
-        merchCard.setAttribute('badge-color', badge.badgeColor);
-        merchCard.setAttribute('badge-text', badge.badgeText);
+  loadStyle(`${base}/deps/caas.css`);
+  const section = el.closest('.section');
+  section.classList.add('milo-card-section');
+  const images = el.querySelectorAll('picture');
+  let image;
+  const icons = [];
+  const rows = el.querySelectorAll(':scope > *');
+  const ribbonMetadata = rows[0].children?.length === 2 ? rows[0].children : null;
+  const row = rows[ribbonMetadata === null ? 0 : 1];
+  const altCta = rows[rows.length - 1].children?.length === 2
+    ? rows[rows.length - 1].children : null;
+  const allPElems = row.querySelectorAll('p');
+  const ctas = allPElems[allPElems.length - 1];
+  const styles = [...el.classList];
+  const cardType = getPodType(styles);
+  const merchCard = el;
+  decorateBlockHrs(el);
+  images.forEach((img) => {
+    const imgNode = img.querySelector('img');
+    const { width, height } = imgNode;
+    const isSquare = Math.abs(width - height) <= 10;
+    if (img) {
+      if (isSquare) {
+        icons.push(img);
+      } else {
+        image = img;
       }
     }
-    if (actionMenuContent !== null) {
-      merchCard.setAttribute('action-menu', true);
-      merchCard.append(createTag('div', { slot: 'action-menu-content' }, actionMenuContent.innerHTML));
-    }
-    if (ctas) {
-      const footer = createTag('div', { slot: 'footer' });
-      decorateButtons(ctas);
-      footer.append(ctas);
-      merchCard.appendChild(footer);
-    }
-    if (image !== undefined) {
-      merchCard.setAttribute('image', image.querySelector('img').src);
-      image.remove();
-    }
-    if (!icons || icons.length > 0) {
-      const iconImgs = Array.from(icons).map((icon) => {
-        const img = {
-          src: icon.querySelector('img').src,
-          alt: icon.querySelector('img').alt,
-        };
-        return img;
-      });
-      merchCard.setAttribute('icons', JSON.stringify(Array.from(iconImgs)));
-      icons.forEach((icon) => icon.remove());
-    }
-    if (styles.includes('secure')) {
-      replaceKey('secure-transaction', getConfig())
-        .then((key) => merchCard.setAttribute('secure-label', key));
-    }
-    if (altCta) {
-      const label = checkBoxLabel(ctas, altCta);
-      if (label !== null) {
-        merchCard.setAttribute('checkbox-label', label);
-      }
-    }
-    parseContent(el, altCta, cardType, merchCard);
-    decorateBlockHrs(merchCard);
-    el.replaceWith(merchCard);
-  }
+  });
+  addWrapper(el, section, cardType);
+  merchCard.classList.add('consonant-Card', 'consonant-ProductCard', `consonant-${cardType}`);
+  if (image) addBackgroundImg(image, cardType, merchCard);
+  if (ribbonMetadata !== null) decorateRibbon(el, ribbonMetadata, cardType);
+  image?.parentElement.remove();
+  if (ctas) decorateButtons(ctas);
+  addInner(el, altCta, cardType, merchCard);
+  decorateIcon(el, icons, cardType);
+  const inner = el.querySelector(`.consonant-${cardType}-inner`);
+  const innerCleanup = inner.querySelectorAll(':scope > div')[1];
+  if (innerCleanup.classList.length === 0) innerCleanup.remove();
 };
 
 export default init;
