--- test/blocks/merch-card/legacy-merch-card/mocks/plans-card.html
@@ -1,76 +1,0 @@
-<div class="section">
-  <div class="merch-card empty">
-    <div>
-      <div></div>
-    </div>
-  </div>
-  <div class="merch-card plans icons secure">
-    <div>
-      <div>#EDCC2D, #000000</div>
-      <div>LOREM IPSUM DOLOR</div>
-    </div>
-    <div>
-      <div>
-        <picture>
-          <source type="image/webp" srcset="" media="(min-width: 600px)">
-          <source type="image/webp" srcset="">
-          <source type="image/png" srcset="" media="(min-width: 600px)">
-          <img loading="lazy" alt="" src="" width="80" height="78">
-        </picture>
-        <picture>
-          <source type="image/webp" srcset="" media="(min-width: 600px)">
-          <source type="image/webp" srcset="">
-          <source type="image/png" srcset="" media="(min-width: 600px)">
-          <img loading="lazy" alt="" src="" width="80" height="78">
-        </picture>
-        <h4 id="lorem-ipsum-dolor-sit-amet"><em>Lorem ipsum dolor sit amet</em></h4>
-        <h3 id="lorem-ipsum-dolor">Lorem ipsum dolor</h3>
-        <p>Maecenas porttitor congue massa. Fusce posuere, magna sed pulvinar ultricies, purus lectus malesuada libero,
-          sit amet commodo magna eros quis urna. Nunc viverra imperdiet enim.</p>
-        <p><strong>Maecenas</strong></p>
-        <ul>
-          <li>Maecenas porttitor congue massa</li>
-          <li>Nunc viverra imperdiet enim.</li>
-        </ul>
-        <p><a href="https://adobe.com/">See terms about lorem ipsum</a></p>
-        <p><em><a href="https://business.adobe.com/">Learn More</a></em> <strong><a href="https://business.adobe.com/">Save now</a></strong></p>
-      </div>
-    </div>
-    <div>
-      <div>Lorem ipsum dolor sit amet</div>
-      <div><a href="/tools/ost?osi=Habitant&#x26;offerId=123456789&#x26;type=checkoutUrl&#x26;text=free-trial">free-trial</a></div>
-    </div>
-  </div>
-  <div class="merch-card plans icons skip-ribbon skip-altCta">
-    <div>
-      <div> </div>
-      <div>LOREM IPSUM DOLOR</div>
-    </div>
-    <div>
-      <div>
-        <picture>
-          <source type="image/webp" srcset="" media="(min-width: 600px)">
-          <source type="image/webp" srcset="">
-          <source type="image/png" srcset="" media="(min-width: 600px)">
-          <img loading="lazy" alt="" src="" width="80" height="78">
-        </picture>
-        <picture>
-          <source type="image/webp" srcset="" media="(min-width: 600px)">
-          <source type="image/webp" srcset="">
-          <source type="image/png" srcset="" media="(min-width: 600px)">
-          <img loading="lazy" alt="" src="" width="80" height="78">
-        </picture>
-        <h4 id="lorem-ipsum-dolor-sit-amet"><em>Lorem ipsum dolor sit amet</em></h4>
-        <h3 id="lorem-ipsum-dolor">Lorem ipsum dolor</h3>
-        <p>Maecenas porttitor congue massa. Fusce posuere, magna sed pulvinar ultricies, purus lectus malesuada libero,
-          sit amet commodo magna eros quis urna. Nunc viverra imperdiet enim.</p>
-        <p><a href="https://adobe.com/">See terms about lorem ipsum</a></p>
-        <p><em><a href="https://business.adobe.com/">Learn More</a></em> <strong><a href="https://business.adobe.com/">Save now</a></strong></p>
-      </div>
-    </div>
-    <div>
-      <div>Lorem ipsum dolor sit amet</div>
-      <div></div>
-    </div>
-  </div>
-</div>
