--- test/blocks/merch-card/legacy-merch-card/mocks/segment-card.html
@@ -1,24 +1,0 @@
-<div class="section">
-  <div class="before"></div>
-  <div class="merch-card segment">
-    <div>
-      <div>
-        <h3 id="lorem-ipsum-dolor-sit-amet">Lorem ipsum dolor sit amet</h3>
-        <p>Maecenas porttitor congue massa. Fusce posuere, magna sed pulvinar ultricies, purus lectus malesuada libero, sit amet commodo magna eros quis urna. Nunc viverra imperdiet enim.</p>
-        <p><a href="https://www.adobe.com/">See what's included</a> | <a href="https://www.adobe.com/">Learn more</a></p>
-        <p><strong><em><a href="https://business.adobe.com/">Learn More</a></em></strong> <strong><a href="https://business.adobe.com/">Save now</a></strong></p>
-      </div>
-    </div>
-  </div>
-  <div class="merch-card segment">
-    <div>
-      <div>
-        <h3 id="lorem-ipsum-dolor-sit-amet">Lorem ipsum dolor sit amet</h3>
-        <p>Maecenas porttitor congue massa. Fusce posuere, magna sed pulvinar ultricies, purus lectus malesuada libero, sit amet commodo magna eros quis urna. Nunc viverra imperdiet enim.</p>
-        <p><a href="https://www.adobe.com/">See what's included</a> | <a href="https://www.adobe.com/">Learn more</a></p>
-        <p><strong><em><a href="https://business.adobe.com/">Learn More</a></em></strong> <strong><a href="https://business.adobe.com/">Save now</a></strong></p>
-      </div>
-    </div>
-  </div>
-  <div class="after"></div>
-</div>
