--- test/blocks/merch-card/legacy-merch-card/mocks/special-offers.html
@@ -1,25 +1,0 @@
-<div class="section">
-  <div class="merch-card special-offers evergreen">
-    <div>
-      <div>#EDCC2D, #000000</div>
-      <div>LOREM IPSUM DOLOR</div>
-    </div>
-    <div>
-      <div>
-        <p>
-          <picture>
-            <source type="image/webp" srcset="" media="(min-width: 600px)">
-            <source type="image/webp" srcset="">
-            <source type="image/png" srcset="" media="(min-width: 600px)">
-            <img loading="lazy" alt="" src="" width="5000" height="2812">
-          </picture>
-        </p>
-        <h4 id="individuals">INDIVIDUALS</h4>
-        <h3 id="get-10-off-photoshop">Get 10% off Photoshop.</h3>
-        <p>Create gorgeous images, rich graphics, and incredible art. Save 10% for the first year. Ends Mar 20.</p>
-        <p><a href="https://adobe.com/">See terms</a></p>
-        <p><em><a href="https://business.adobe.com/">Learn More</a></em> <strong><a href="https://business.adobe.com/">Save now</a></strong></p>
-      </div>
-    </div>
-  </div>
-</div>
