--- test/blocks/merch-card/legacy-merch-card/mocks/uar-card.html
@@ -1,49 +1,0 @@
-<div class="section">
-  
-  <h1>UAR CARDS</h1>
-  
-  <div class="merch-card">
-    <div>
-      <div data-valign="middle">
-        <p><a href="/drafts/rparrish/assets/merch/photoshop.svg">https://main--milo--adobecom.hlx.page/drafts/rparrish/assets/merch/photoshop.svg</a></p>
-        <h4 id="photoshop">Photoshop</h4>
-        <p><a href="/tools/ost?osi=2maG5a9vLOf7LjkcfwPbq-JT7RK8ZJ8RpyGaVc4DRQQ&amp;offerId=28A8BD2F1A07795E4C5C37E59BA351B1&amp;type=priceOptical&amp;term=true&amp;seat=false&amp;tax=false">PRICE - PUF - Photoshop with 2TB</a> <a href="/tools/ost?osi=41wh2RRZK4M4qTbbEhzF2Xe9YicCIrNCfQQuxveECTg&amp;offerId=D090AD1C59E7360A2880EBFE339F6BB6&amp;type=priceStrikethrough&amp;term=true&amp;seat=false&amp;tax=false">PRICE - ABM - Photoshop with 2TB</a></p>
-        <p>HR in between price and link? Variant UAR - Create gorgeous images, rich graphics, and incredible art.</p>
-        <p><br>--- #00ffff</p>
-        <p><a href="./">See terms</a></p>
-        <p><em><a href="./">Learn More</a></em> <strong><a href="./">Save now</a></strong></p>
-      </div>
-    </div>
-  </div>
-
-  <h2>inline heading</h2>
-  <div class="merch-card inline-heading">
-    <div>
-      <div data-valign="middle">
-        <h2 id="httpsmain--milo--adobecomhlxpagedraftsrparrishassetsmerchindesignsvg"><a href="/drafts/rparrish/assets/merch/indesign.svg">https://main--milo--adobecom.hlx.page/drafts/rparrish/assets/merch/indesign.svg</a></h2>
-        <h4 id="indesign">InDesign</h4>
-        <p><a href="/tools/ost?osi=2maG5a9vLOf7LjkcfwPbq-JT7RK8ZJ8RpyGaVc4DRQQ&amp;offerId=28A8BD2F1A07795E4C5C37E59BA351B1&amp;type=priceOptical&amp;term=true&amp;seat=false&amp;tax=false">PRICE - PUF - Photoshop with 2TB</a> <a href="/tools/ost?osi=41wh2RRZK4M4qTbbEhzF2Xe9YicCIrNCfQQuxveECTg&amp;offerId=D090AD1C59E7360A2880EBFE339F6BB6&amp;type=priceStrikethrough&amp;term=true&amp;seat=false&amp;tax=false">PRICE - ABM - Photoshop with 2TB</a></p>
-        <p>--- #FF3266</p>
-        <p>Variant: ‘inline headline’ is used to keep the title inline w/ the icon.</p>
-        <p><a href="./">See terms</a></p>
-        <p><em><a href="./">Learn More</a></em> <strong><a href="./">Save now</a></strong></p>
-      </div>
-    </div>
-  </div>
-
-  <h2>background opacity 70</h2>
-  <div class="merch-card background-opacity-70">
-    <div>
-      <div data-valign="middle">
-        <p><a href="/drafts/rparrish/assets/merch/xd.svg">https://main--milo--adobecom.hlx.page/drafts/rparrish/assets/merch/xd.svg</a></p>
-        <h4 id="xd">XD</h4>
-        <h4 id="price---puf---photoshop-with-2tb-price---abm---photoshop-with-2tb"><a href="/tools/ost?osi=2maG5a9vLOf7LjkcfwPbq-JT7RK8ZJ8RpyGaVc4DRQQ&amp;offerId=28A8BD2F1A07795E4C5C37E59BA351B1&amp;type=priceOptical&amp;term=true&amp;seat=false&amp;tax=false">PRICE - PUF - Photoshop with 2TB</a> <a href="/tools/ost?osi=41wh2RRZK4M4qTbbEhzF2Xe9YicCIrNCfQQuxveECTg&amp;offerId=D090AD1C59E7360A2880EBFE339F6BB6&amp;type=priceStrikethrough&amp;term=true&amp;seat=false&amp;tax=false">PRICE - ABM - Photoshop with 2TB</a></h4>
-        <p>HR between link and ctas? Variant: ‘background opacity 70’ is used to have a 70% white bg.</p>
-        <p><a href="https://adobe.com/">See terms</a></p>
-        <p><a href="./">See terms</a></p>
-        <p><em><a href="./">Learn More</a></em> <strong><a href="./">Save now</a></strong></p>
-      </div>
-    </div>
-  </div>
-
-</div>
