--- test/blocks/merch-card/mocks/catalog.html
@@ -1,109 +1,0 @@
-<div class="section">
-  <div class="merch-card catalog ribbon">
-    <div>
-      <div>#EDCC2D, #000000</div>
-      <div>LOREM IPSUM DOLOR</div>
-    </div>
-    <div>
-      <div data-valign="middle">
-        <p><strong>Best for</strong>:</p>
-        <ul>
-          <li>Photo editing</li>
-          <li>Compositing</li>
-          <li>Drawing and painting</li>
-          <li>Graphic design</li>
-        </ul>
-        <p><strong>Storage</strong>:</p>
-        <p>100 GB of cloud storage</p>
-        <p><a href="https://adobe.com">See system requirements</a></p>
-      </div>
-    </div>
-    <div>
-      <div>
-        <picture>
-          <source type="image/webp" srcset="" media="(min-width: 600px)">
-          <source type="image/webp" srcset="">
-          <source type="image/png" srcset="" media="(min-width: 600px)">
-          <img loading="lazy" alt="" src="" width="5000" height="2812">
-        </picture>
-        <h4 id="individuals">INDIVIDUALS</h4>
-        <h3 id="get-10-off-photoshop">Get 10% off Photoshop.</h3>
-        <h5>Desktop + Mobile</h5>
-        <p>Create gorgeous images, rich graphics, and incredible art. Save 10% for the first year. Ends Mar 20.</p>
-        <p><a href="https://adobe.com/">See terms</a></p>
-        <p><em><a href="https://business.adobe.com/">Learn More</a></em> <strong><a href="https://business.adobe.com/">Save now</a></strong></p>
-      </div>
-    </div>
-  </div>
-  <div class="merch-card catalog empty-badge">
-    <div data-valign="middle">
-      <p><strong>Best for</strong>:</p>
-      <ul>
-        <li>Photo editing</li>
-        <li>Compositing</li>
-        <li>Drawing and painting</li>
-        <li>Graphic design</li>
-      </ul>
-      <p><strong>Storage</strong>:</p>
-      <p>100 GB of cloud storage</p>
-      <p><a href="https://adobe.com">See system requirements</a></p>
-    </div>
-    <div>
-      <div>
-        <picture>
-          <source type="image/webp" srcset="" media="(min-width: 600px)">
-          <source type="image/webp" srcset="">
-          <source type="image/png" srcset="" media="(min-width: 600px)">
-          <img loading="lazy" alt="" src="" width="5000" height="2812">
-        </picture>
-        <h1 id="individuals1">INDIVIDUALS</h1>
-        <h2 id="get-10-off-photoshop1">Get 10% off Photoshop.</h2>
-        <h5>Desktop + Mobile</h5>
-        <p>Create gorgeous images, rich graphics, and incredible art. Save 10% for the first year. Ends Mar 20.</p>
-        <p><a href="https://adobe.com/">See terms</a></p>
-        <p><em><a href="https://business.adobe.com/">Learn More</a></em> <strong><a href="https://business.adobe.com/">Save now</a></strong></p>
-      </div>
-    </div>
-  </div>
-  <div class="merch-card catalog empty-action-menu">
-    <div>
-      <div>
-        <picture>
-          <source type="image/webp" srcset="" media="(min-width: 600px)">
-          <source type="image/webp" srcset="">
-          <source type="image/png" srcset="" media="(min-width: 600px)">
-          <img loading="lazy" alt="" src="" width="5000" height="2812">
-        </picture>
-        <h1 id="individuals2">INDIVIDUALS</h1>
-        <h2 id="get-10-off-photoshop2">Get 10% off Photoshop.</h2>
-        <h5>Desktop + Mobile</h5>
-        <p>Create gorgeous images, rich graphics, and incredible art. Save 10% for the first year. Ends Mar 20.</p>
-        <p><a href="https://adobe.com/">See terms</a></p>
-        <p><em><a href="https://business.adobe.com/">Learn More</a></em> <strong><a
-          href="https://business.adobe.com/">Save now</a></strong></p>
-      </div>
-    </div>
-  </div>
-  <div class="merch-card catalog empty-badge action-menu-exist">
-    <div>
-      <div>#EDCC2D, #000000</div>
-      <div>LOREM IPSUM DOLOR</div>
-    </div>
-    <div>
-      <div>
-        <picture>
-          <source type="image/webp" srcset="" media="(min-width: 600px)">
-          <source type="image/webp" srcset="">
-          <source type="image/png" srcset="" media="(min-width: 600px)">
-          <img loading="lazy" alt="" src="" width="5000" height="2812">
-        </picture>
-        <h2 id="individuals1">INDIVIDUALS</h2>
-        <h3 id="get-10-off-photoshop1">Get 10% off Photoshop.</h3>
-        <h5>Desktop + Mobile</h5>
-        <p>Create gorgeous images, rich graphics, and incredible art. Save 10% for the first year. Ends Mar 20.</p>
-        <p><a href="https://adobe.com/">See terms</a></p>
-        <p><em><a href="https://business.adobe.com/">Learn More</a></em> <strong><a href="https://business.adobe.com/">Save now</a></strong></p>
-      </div>
-    </div>
-  </div>
-</div>
