+++ test/blocks/merch-card/mocks/plans-card.html
@@ -23,12 +23,15 @@
           <source type="image/png" srcset="" media="(min-width: 600px)">
           <img loading="lazy" alt="" src="" width="80" height="78">
         </picture>
-        <h3 id="lorem-ipsum-dolor-sit-amet1">Lorem ipsum dolor sit amet</h3>
-        <h2 id="lorem-ipsum-dolor1">Lorem ipsum dolor</h2>
-        <h4 id="lorem-ipsum-dolor-sit">Lorem ipsum dolor sit</h4>
-        <h5>Maecenas porttitor enim.</h5>
-        <p>Maecenas porttitor congue massa. Fusce posuere, magna sed pulvinar ultricies, purus lectus malesuada libero, sit amet commodo magna eros quis urna. Nunc viverra imperdiet enim.</p>
+        <h4 id="lorem-ipsum-dolor-sit-amet"><em>Lorem ipsum dolor sit amet</em></h4>
+        <h3 id="lorem-ipsum-dolor">Lorem ipsum dolor</h3>
+        <p>Maecenas porttitor congue massa. Fusce posuere, magna sed pulvinar ultricies, purus lectus malesuada libero,
+          sit amet commodo magna eros quis urna. Nunc viverra imperdiet enim.</p>
         <p><strong>Maecenas</strong></p>
+        <ul>
+          <li>Maecenas porttitor congue massa</li>
+          <li>Nunc viverra imperdiet enim.</li>
+        </ul>
         <p><a href="https://adobe.com/">See terms about lorem ipsum</a></p>
         <p><em><a href="https://business.adobe.com/">Learn More</a></em> <strong><a href="https://business.adobe.com/">Save now</a></strong></p>
       </div>
@@ -57,10 +60,10 @@
           <source type="image/png" srcset="" media="(min-width: 600px)">
           <img loading="lazy" alt="" src="" width="80" height="78">
         </picture>
-        <h3 id="lorem-ipsum-dolor-sit-amet">Lorem ipsum dolor sit lorem-ipsum-dolor-sit-amet1</h3>
-        <h2 id="lorem-ipsum-dolor">Lorem ipsum dolor</h2>
-        <h4 id="lorem-ipsum-dolor-sit">Lorem ipsum dolor sit</h4>
-        <p>Maecenas porttitor congue massa. Fusce posuere, magna sed pulvinar ultricies, purus lectus malesuada libero, sit amet commodo magna eros quis urna. Nunc viverra imperdiet enim.</p>
+        <h4 id="lorem-ipsum-dolor-sit-amet"><em>Lorem ipsum dolor sit amet</em></h4>
+        <h3 id="lorem-ipsum-dolor">Lorem ipsum dolor</h3>
+        <p>Maecenas porttitor congue massa. Fusce posuere, magna sed pulvinar ultricies, purus lectus malesuada libero,
+          sit amet commodo magna eros quis urna. Nunc viverra imperdiet enim.</p>
         <p><a href="https://adobe.com/">See terms about lorem ipsum</a></p>
         <p><em><a href="https://business.adobe.com/">Learn More</a></em> <strong><a href="https://business.adobe.com/">Save now</a></strong></p>
       </div>
